Refactoring to use trigger task
