+++ src/integrations/prefect-dbt/tests/cli/test_commands.py
@@ -4,60 +4,114 @@ from unittest.mock import MagicMock
 
 import pytest
 import yaml
-from prefect_dbt.cli.commands import DbtCoreOperation, trigger_dbt_cli_command
+from dbt.cli.main import dbtRunnerResult
+from dbt.contracts.files import FileHash
+from dbt.contracts.graph.nodes import ModelNode
+from dbt.contracts.results import RunExecutionResult, RunResult
+from prefect_dbt.cli.commands import (
+    DbtCoreOperation,
+    run_dbt_build,
+    run_dbt_model,
+    run_dbt_seed,
+    run_dbt_snapshot,
+    run_dbt_test,
+    trigger_dbt_cli_command,
+)
 from prefect_dbt.cli.credentials import DbtCliProfile
 
 from prefect import flow
+from prefect.artifacts import Artifact
 from prefect.testing.utilities import AsyncMock
 
 
-async def mock_shell_run_command_fn(**kwargs):
-    return kwargs
-
-
-@pytest.fixture(autouse=True)
-def mock_shell_run_command(monkeypatch):
-    _mock_shell_run_command = MagicMock(fn=mock_shell_run_command_fn)
-    monkeypatch.setattr(
-        "prefect_dbt.cli.commands.shell_run_command", _mock_shell_run_command
+@pytest.fixture
+async def mock_dbt_runner_model_success():
+    return dbtRunnerResult(
+        success=True,
+        exception=None,
+        result=RunExecutionResult(
+            results=[
+                RunResult(
+                    status="pass",
+                    timing=None,
+                    thread_id="'Thread-1 (worker)'",
+                    message="CREATE TABLE (1.0 rows, 0 processed)",
+                    failures=None,
+                    node=ModelNode(
+                        database="test-123",
+                        schema="prefect_dbt_example",
+                        name="my_first_dbt_model",
+                        resource_type="model",
+                        package_name="prefect_dbt_bigquery",
+                        path="example/my_first_dbt_model.sql",
+                        original_file_path="models/example/my_first_dbt_model.sql",
+                        unique_id="model.prefect_dbt_bigquery.my_first_dbt_model",
+                        fqn=["prefect_dbt_bigquery", "example", "my_first_dbt_model"],
+                        alias="my_first_dbt_model",
+                        checksum=FileHash(name="sha256", checksum="123456789"),
+                    ),
+                    execution_time=0.0,
+                    adapter_response=None,
+                )
+            ],
+            elapsed_time=0.0,
+        ),
     )
 
 
 @pytest.fixture
+async def mock_dbt_runner_ls_success():
+    return dbtRunnerResult(
+        success=True, exception=None, result=["example.example.test_model"]
+    )
+
+
+@pytest.fixture
+def dbt_runner_model_result(monkeypatch, mock_dbt_runner_model_success):
+    _mock_dbt_runner_invoke_success = MagicMock(
+        return_value=mock_dbt_runner_model_success
+    )
+    monkeypatch.setattr(
+        "dbt.cli.main.dbtRunner.invoke", _mock_dbt_runner_invoke_success
+    )
+
+
+@pytest.fixture
+def dbt_runner_ls_result(monkeypatch, mock_dbt_runner_ls_success):
+    _mock_dbt_runner_ls_result = MagicMock(return_value=mock_dbt_runner_ls_success)
+    monkeypatch.setattr("dbt.cli.main.dbtRunner.invoke", _mock_dbt_runner_ls_result)
+
+
+@pytest.fixture
 def profiles_dir(tmp_path):
-    return tmp_path / ".dbt"
+    return str(tmp_path) + "/.dbt"
 
 
-def test_trigger_dbt_cli_command_not_dbt():
-    @flow
-    def test_flow():
-        return trigger_dbt_cli_command("ls")
-
-    with pytest.raises(ValueError, match="Command is not a valid dbt sub-command"):
-        test_flow()
-
-
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command(profiles_dir, dbt_cli_profile_bare):
     @flow
     def test_flow():
         return trigger_dbt_cli_command(
-            "dbt ls", profiles_dir=profiles_dir, dbt_cli_profile=dbt_cli_profile_bare
+            command="ls",
+            profiles_dir=profiles_dir,
+            dbt_cli_profile=dbt_cli_profile_bare,
         )
 
     result = test_flow()
-    assert result == {"command": f"dbt ls --profiles-dir {profiles_dir}"}
+    assert isinstance(result, dbtRunnerResult)
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_run_twice_overwrite(
     profiles_dir, dbt_cli_profile, dbt_cli_profile_bare
 ):
     @flow
     def test_flow():
         trigger_dbt_cli_command(
-            "dbt ls", profiles_dir=profiles_dir, dbt_cli_profile=dbt_cli_profile
+            command="ls", profiles_dir=profiles_dir, dbt_cli_profile=dbt_cli_profile
         )
         run_two = trigger_dbt_cli_command(
-            "dbt ls",
+            command="ls",
             profiles_dir=profiles_dir,
             dbt_cli_profile=dbt_cli_profile_bare,
             overwrite_profiles=True,
@@ -65,8 +119,8 @@ def test_trigger_dbt_cli_command_run_twice_overwrite(
         return run_two
 
     result = test_flow()
-    assert result == {"command": f"dbt ls --profiles-dir {profiles_dir}"}
-    with open(profiles_dir / "profiles.yml", "r") as f:
+    assert isinstance(result, dbtRunnerResult)
+    with open(profiles_dir + "/profiles.yml", "r") as f:
         actual = yaml.safe_load(f)
     expected = {
         "config": {},
@@ -85,16 +139,17 @@ def test_trigger_dbt_cli_command_run_twice_overwrite(
     assert actual == expected
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_run_twice_exists(
     profiles_dir, dbt_cli_profile, dbt_cli_profile_bare
 ):
     @flow
     def test_flow():
         trigger_dbt_cli_command(
-            "dbt ls", profiles_dir=profiles_dir, dbt_cli_profile=dbt_cli_profile
+            "ls", profiles_dir=profiles_dir, dbt_cli_profile=dbt_cli_profile
         )
         run_two = trigger_dbt_cli_command(
-            "dbt ls",
+            "ls",
             profiles_dir=profiles_dir,
             dbt_cli_profile=dbt_cli_profile_bare,
         )
@@ -104,6 +159,7 @@ def test_trigger_dbt_cli_command_run_twice_exists(
         test_flow()
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_missing_profile(profiles_dir):
     @flow
     def test_flow():
@@ -118,6 +174,7 @@ def test_trigger_dbt_cli_command_missing_profile(profiles_dir):
         test_flow()
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_find_home(dbt_cli_profile_bare):
     home_dbt_dir = Path.home() / ".dbt"
     if (home_dbt_dir / "profiles.yml").exists():
@@ -128,23 +185,25 @@ def test_trigger_dbt_cli_command_find_home(dbt_cli_profile_bare):
     @flow
     def test_flow():
         return trigger_dbt_cli_command(
-            "dbt ls", dbt_cli_profile=dbt_cli_profile, overwrite_profiles=False
+            "ls", dbt_cli_profile=dbt_cli_profile, overwrite_profiles=False
         )
 
     result = test_flow()
-    assert result == {"command": f"dbt ls --profiles-dir {home_dbt_dir}"}
+    assert isinstance(result, dbtRunnerResult)
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_find_env(profiles_dir, dbt_cli_profile_bare):
     @flow
     def test_flow():
-        return trigger_dbt_cli_command("dbt ls", dbt_cli_profile=dbt_cli_profile_bare)
+        return trigger_dbt_cli_command("ls", dbt_cli_profile=dbt_cli_profile_bare)
 
     os.environ["DBT_PROFILES_DIR"] = str(profiles_dir)
     result = test_flow()
-    assert result == {"command": f"dbt ls --profiles-dir {profiles_dir}"}
+    assert isinstance(result, dbtRunnerResult)
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_project_dir(profiles_dir, dbt_cli_profile_bare):
     @flow
     def test_flow():
@@ -156,11 +215,10 @@ def test_trigger_dbt_cli_command_project_dir(profiles_dir, dbt_cli_profile_bare)
         )
 
     result = test_flow()
-    assert result == {
-        "command": f"dbt ls --profiles-dir {profiles_dir} --project-dir project"
-    }
+    assert isinstance(result, dbtRunnerResult)
 
 
+@pytest.mark.usefixtures("dbt_runner_ls_result")
 def test_trigger_dbt_cli_command_shell_kwargs(profiles_dir, dbt_cli_profile_bare):
     @flow
     def test_flow():
@@ -172,10 +230,7 @@ def test_trigger_dbt_cli_command_shell_kwargs(profiles_dir, dbt_cli_profile_bare
         )
 
     result = test_flow()
-    assert result == {
-        "command": f"dbt ls --profiles-dir {profiles_dir}",
-        "return_all": True,
-    }
+    assert isinstance(result, dbtRunnerResult)
 
 
 class TestDbtCoreOperation:
@@ -288,3 +343,93 @@ class TestDbtCoreOperation:
             mock_write.call_args_list[0][0][0]
             == f"dbt debug --profiles-dir {tmp_path} --project-dir {tmp_path}".encode()
         )
+
+
+@pytest.mark.usefixtures("dbt_runner_model_result")
+def test_run_dbt_build_creates_artifact(profiles_dir, dbt_cli_profile_bare):
+    @flow
+    def test_flow():
+        return run_dbt_build(
+            profiles_dir=profiles_dir,
+            dbt_cli_profile=dbt_cli_profile_bare,
+            artifact_key="foo",
+            create_artifact=True,
+        )
+
+    test_flow()
+    assert (a := Artifact.get(key="foo"))
+    assert a.type == "markdown"
+    assert a.data.startswith("# dbt build Task Summary")
+    assert "my_first_dbt_model" in a.data
+
+
+@pytest.mark.usefixtures("dbt_runner_model_result")
+def test_run_dbt_test_creates_artifact(profiles_dir, dbt_cli_profile_bare):
+    @flow
+    def test_flow():
+        return run_dbt_test(
+            profiles_dir=profiles_dir,
+            dbt_cli_profile=dbt_cli_profile_bare,
+            artifact_key="foo",
+            create_artifact=True,
+        )
+
+    test_flow()
+    assert (a := Artifact.get(key="foo"))
+    assert a.type == "markdown"
+    assert a.data.startswith("# dbt test Task Summary")
+    assert "my_first_dbt_model" in a.data
+
+
+@pytest.mark.usefixtures("dbt_runner_model_result")
+def test_run_dbt_snapshot_creates_artifact(profiles_dir, dbt_cli_profile_bare):
+    @flow
+    def test_flow():
+        return run_dbt_snapshot(
+            profiles_dir=profiles_dir,
+            dbt_cli_profile=dbt_cli_profile_bare,
+            artifact_key="foo",
+            create_artifact=True,
+        )
+
+    test_flow()
+    assert (a := Artifact.get(key="foo"))
+    assert a.type == "markdown"
+    assert a.data.startswith("# dbt snapshot Task Summary")
+    assert "my_first_dbt_model" in a.data
+
+
+@pytest.mark.usefixtures("dbt_runner_model_result")
+def test_run_dbt_seed_creates_artifact(profiles_dir, dbt_cli_profile_bare):
+    @flow
+    def test_flow():
+        return run_dbt_seed(
+            profiles_dir=profiles_dir,
+            dbt_cli_profile=dbt_cli_profile_bare,
+            artifact_key="foo",
+            create_artifact=True,
+        )
+
+    test_flow()
+    assert (a := Artifact.get(key="foo"))
+    assert a.type == "markdown"
+    assert a.data.startswith("# dbt seed Task Summary")
+    assert "my_first_dbt_model" in a.data
+
+
+@pytest.mark.usefixtures("dbt_runner_model_result")
+def test_rub_dbt_model_creates_artifact(profiles_dir, dbt_cli_profile_bare):
+    @flow
+    def test_flow():
+        return run_dbt_model(
+            profiles_dir=profiles_dir,
+            dbt_cli_profile=dbt_cli_profile_bare,
+            artifact_key="foo",
+            create_artifact=True,
+        )
+
+    test_flow()
+    assert (a := Artifact.get(key="foo"))
+    assert a.type == "markdown"
+    assert a.data.startswith("# dbt run Task Summary")
+    assert "my_first_dbt_model" in a.data
